requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
attrs==25.3.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
chardet==5.2.0
charset-normalizer==2.0.12
click==8.1.8
cryptography==41.0.5
cssselect2==0.7.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.11
fonttools==4.56.0
frozenlist==1.5.0
google-api-core==2.24.1
google-auth==2.38.0
google-maps-routeoptimization==0.1.8
googleapis-common-protos==1.69.0
grpcio==1.71.0rc2
grpcio-status==1.71.0rc2
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.2.0
pillow==11.1.0
propcache==0.3.1
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydyf==0.11.0
pyphen==0.17.2
python-dotenv==0.19.0
reportlab==4.3.1
requests==2.26.0
rsa==4.9
setuptools==75.8.2
sniffio==1.3.1
starlette==0.46.0
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
urllib3==1.26.20
uvicorn==0.34.0
weasyprint==64.1
webencodings==0.5.1
wheel==0.45.1
yarl==1.18.3
zopfli==0.2.3.post1
